First blockout test
Just testing some stuff out, not definite or final by any means. We discussed an enclosed area with some visibility of the surrounding areas

- Removed test assets from earlier
// Deleted from scene, they're still in the project files

- Made a simple fence prefab
// Can be used to mark the playable area for the exterior part.

- First test of scale.
// Main building will probably need to be bigger and a more interesting shape, but hard to say without a player controller.

- Road planes
// Added some planes for roads, to see how it might look to have streets next to the facility compound area. Feel free to start adding blocks in the background to see how much is needed to fill in the urban surroundings.

Need a basic character controller ASAP, to be able to get a better sense of scale and so we can test layouts, object placements, movement paths etc.