New Third Person Camera & Jumping Animation
Implemented new TPS camera using Cinemachine.
Added a jumping animation.

Jump animation is a bit scuffed, same as the others, but the main issue I'm seeing is that the player script doesn't ever reset to "IsGrounded". It's on at start of play, and is checked for in the Jump script, but there's nothing to set it back once the player has landed.