Created & added Healthbar UI to Canvas + connected it to Health script
Using LeftMouse click to take damage, just for testing the healthbar functionality.

Will connect to turret projectile next, and should be easy to add any other potential damage sources, as well as healing if we want something like that.

Oh also, added the Unity 2D package to the project just to have access to more sprites and stuff to use in UI mockups.